refactor(flipCard): Expose a cleaner context with safe typing
