--- src/pages/Campaign/widgets/widgetCards/FlipCard/FlipCardContext.ts
@@ -1,4 +1,0 @@
-import { createContext } from 'react';
-import { FlipCardContextType } from './types';
-
-export const FlipCardContext = createContext({} as FlipCardContextType);
