+++ src/pages/Campaign/widgets/widgetCards/FlipCard/context/FlipCardContext.tsx
@@ -1,0 +1,35 @@
+import { createContext, useContext, useMemo, useState } from 'react';
+import { FaceType, FlipCardContextType } from '../types';
+
+export const FlipCardContext = createContext<FlipCardContextType | null>(null);
+
+export const FlipCardContextProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [visibleFace, setVisibleFace] = useState<FaceType>('front');
+
+  const flipCardContextValue = useMemo(
+    () => ({
+      visibleFace,
+      setVisibleFace,
+    }),
+    [visibleFace, setVisibleFace]
+  );
+
+  return (
+    <FlipCardContext.Provider value={flipCardContextValue}>
+      {children}
+    </FlipCardContext.Provider>
+  );
+};
+
+export const useFlipCardContext = () => {
+  const context = useContext(FlipCardContext);
+
+  if (!context)
+    throw new Error('Provider not found for FlipCardContextProvider');
+
+  return context; // Now we can use the context in the component, SAFELY.
+};
